added docu folder and added more infos to kdoc. removed private repository declaration

+++ lowlevel-rscp/src/main/kotlin/de/jnkconsulting/e3dc/easyrscp/connection/DefaultSocketFactory.kt
@@ -15,6 +15,6 @@ class DefaultSocketFactory: SocketFactory {
     override fun createSocket(connectionData: E3DCConnectionData) =
         Socket(connectionData.address, connectionData.port).apply {
             tcpNoDelay = true
-            soTimeout = connectionData.timeoutMillis
+            soTimeout = connectionData.timeoutMillis.toInt()
         }
 }
